move non-test files out of tests/
